Merges InternalLibraryConventionPlugin and PublicLibraryConventionPlugin into RevenueCatLibraryConventionPlugin
